First, confire the ApplicationDbContext and include the Entities. For this project, the User entitie was modificated in his primary key, because it is a string type by default. For more information visit: https://docs.microsoft.com/en-us/aspnet/identity/overview/extensibility/change-primary-key-for-users-in-aspnet-identity
Next, was developed the Register funtionality. This one, works in the common area that .Net make by default, but the process to creation an User is different because the project has a Person table. See PersonViewModel, PersonController and AccountController.

Additionaly, the project counts with ambiance.js for magane the notifications